+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Pending
+
+* Make it possible to focus on a hierarchy / tree of tests using `focused` instead of `to` on context level.
+  See [Focusing tests](docs/running-tests.md#Focusing-tests) for more details.
+
+* Make it possible to ignore on a hierarchy / tree of tests using `ignored` instead of `to` on context level.
+  See [Focusing tests](docs/running-tests.md#Focusing-tests) for more details.
+
 ## 0.1.0 (2019-10-17)
 
 * Initial public release